requirements.txt
streamlit>=1.40.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
pandas>=2.2.0
python-dotenv>=1.0.0
